pyproject.toml
[project]
name = "objectdetection"
version = "0.1.0"
description = "This is a sample product or projecrt for Object Detection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.114.2",
    "strawberry-graphql[fastapi]>=0.240.4",
    "ultralytics>=8.2.93",
    "uvicorn>=0.30.6",
]
